requirements.txt
asyncpg
discord.py
flake8
fastapi
